autoplacement: default to use separate machines
Previously the default was to combine everything onto a single system
and use LXC unless otherwise overridden in the charm metadata.

This still doesn't pay attention to the allowed_assignment_types in the
autoplacement, but it makes that less likely to be a problem. In the
openstack installer that this code came from, the only charms which
didn't allow LXC install types were also set with isolate=True, so they
didn't hit this either.

Signed-off-by: Michael McCracken <mike.mccracken@canonical.com>