implement charm equality
Much of the controller code assumes that the charm classes are actually
python classes. For simplicity in creating them dynamically we've made
them all objects of a single Charm class. So we need to implement __eq__
so existing code that compares many separate objects created for the
same charm can work without changing.

Signed-off-by: Michael McCracken <mike.mccracken@canonical.com>