use subordinate buttons in service widgets
make simpleservicewidgets do a similar multi-state button thing, where
clicking once shows you options for what to do next.

The options include placement editing - the default - or relation
editing.

Relation editing is not implemented fully in this commit, it'll be based
on the RelationsColumn class, which is currently just hacked up
copypasta from an earlier design.

Signed-off-by: Michael McCracken <mike.mccracken@canonical.com>