Refactor, separate charms and services
the Charm class and charm_class convention were leftovers from a
previous version of the codebase that didn't need to worry about service
names not matching the charm. This was confusing.

Now all bundle services are Service classes, and most things are keyed
by service instead of charm - including relations, which fixes a bug
prior to this commit that had us failing to correctly detect that a
relation was already connected in a bundle, because the bundle relations
are between services, not charms.

Signed-off-by: Michael McCracken <mike.mccracken@canonical.com>